Remove legacy canvas system and add gallery modal/content components
Delete the unused draggable canvas system (components + canvasData) and
introduce GalleryContent and GalleryModal, updating gallery and link-library
components for the new gallery-based layout.
