--- src/components/canvas/CardContent.jsx
@@ -1,114 +1,0 @@
-'use client';
-
-import { memo, Suspense } from 'react';
-import { Play } from 'lucide-react';
-import { useState } from 'react';
-import dynamic from 'next/dynamic';
-import CurvedArrow from './arrows/CurvedArrow';
-import { ErrorBoundary } from '@/components/ErrorBoundary';
-
-// Loading skeleton for dynamic components
-const ComponentSkeleton = () => (
-  <div className="w-full h-full flex items-center justify-center bg-slate-100 dark:bg-slate-800 rounded-lg animate-pulse">
-    <div className="w-8 h-8 border-4 border-slate-300 dark:border-slate-600 border-t-slate-600 dark:border-t-slate-300 rounded-full animate-spin" />
-  </div>
-);
-
-// Component registry
-const componentMap = {
-  Clock: dynamic(() => import('@/app/experiences/clock/Clock')),
-  ClockPreview: dynamic(() => import('@/components/previews/ClockPreview')),
-  BreathingHero: dynamic(() => import('@/components/effects/BreathingHero')),
-  CoinFlip: dynamic(() => import('@/app/svg-animations/coinflip/CoinFlip')),
-};
-
-function CardContent({ contentType, content, component, componentProps, title }) {
-  const [isPlaying, setIsPlaying] = useState(false);
-
-  // Image content
-  if (contentType === 'image') {
-    return (
-      <div className="relative w-full h-full">
-        <img
-          src={content}
-          alt={title}
-          className="w-full h-full object-cover rounded-lg"
-        />
-      </div>
-    );
-  }
-
-  // Video content
-  if (contentType === 'video') {
-    return (
-      <div className="relative w-full h-full rounded-lg overflow-hidden">
-        <video
-          src={content}
-          className="w-full h-full object-cover rounded-lg"
-          autoPlay
-          muted
-          loop
-          playsInline
-        />
-      </div>
-    );
-  }
-
-  // Text content
-  if (contentType === 'text') {
-    return (
-      <div className="w-full h-full flex items-center justify-center">
-        <p className="text-lg text-slate-700 dark:text-slate-300 text-center whitespace-pre-line font-medium">
-          {content}
-        </p>
-      </div>
-    );
-  }
-
-  // Group title
-  if (contentType === 'group-title') {
-    return (
-      <div className="w-full h-full flex items-center justify-center">
-        <h2 className="font-medium text-orange-900 dark:text-slate-100 font-heading">
-          {content}
-        </h2>
-      </div>
-    );
-  }
-
-  // Arrow
-  if (contentType === 'arrow') {
-    return (
-      <div className="w-full h-full flex items-center justify-center">
-        <CurvedArrow direction={content} width="100%" height="100%" />
-      </div>
-    );
-  }
-
-  // Component content
-  if (contentType === 'component') {
-    const Component = componentMap[component];
-
-    if (!Component) {
-      return (
-        <div className="w-full h-full flex items-center justify-center bg-gray-100 rounded-lg">
-          <p className="text-sm text-gray-500">Component not found</p>
-        </div>
-      );
-    }
-
-    return (
-      <div className="w-full h-full flex items-center justify-center" suppressHydrationWarning>
-        <ErrorBoundary>
-          <Suspense fallback={<ComponentSkeleton />}>
-            <Component {...(componentProps || {})} />
-          </Suspense>
-        </ErrorBoundary>
-      </div>
-    );
-  }
-
-  return null;
-}
-
-export default memo(CardContent);
