--- src/components/canvas/arrows/CurvedArrow.jsx
@@ -1,46 +1,0 @@
-export default function CurvedArrow({ direction = 'down-left', width = 100, height = 100 }) {
-  const arrows = {
-    'spec-left': (
-      <svg viewBox="0 0 1282 547" className="w-full h-full" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M1267 15C1222.05 16.926 1172.95 37.0566 1130.3 51.8763C1080.91 69.0343 1039 83.4244 1006.02 126.431C980.707 159.442 968.128 188.771 971.543 231.448C975.59 282.033 1009.11 326.778 1015.24 376.949C1021.06 424.526 1021.46 464.036 985.173 498.801C928.781 552.835 830.709 527.177 763.28 516.438C639.461 496.717 516.163 473.76 392.455 453.307C293.811 436.998 198.237 419.036 98 419.036" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-        <path d="M105.787 346.787C118.2 329.265 118.758 335.871 105.787 346.987C83.9687 365.686 54.2417 373.163 32.3714 391.791C11.7474 409.357 7.32849 423.084 30.9711 439.594C74.5779 470.046 126.863 478.955 167 516" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-      </svg>
-    ),
-    'spec-right': (
-      <svg viewBox="0 0 1282 547" className="w-full h-full" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M15 15C59.9481 16.926 109.045 37.0566 151.704 51.8763C201.093 69.0343 243.004 83.4244 275.98 126.431C301.293 159.442 313.872 188.771 310.457 231.448C306.41 282.033 272.893 326.778 266.76 376.949C260.944 424.526 260.544 464.036 296.827 498.801C353.219 552.835 451.291 527.177 518.72 516.438C642.539 496.717 765.837 473.76 889.545 453.307C988.189 436.998 1083.76 419.036 1184 419.036" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-        <path d="M1176.21 346.787C1163.8 329.265 1163.24 335.871 1176.21 346.987C1198.03 365.686 1227.76 373.163 1249.63 391.791C1270.25 409.357 1274.67 423.084 1251.03 439.594C1207.42 470.046 1155.14 478.955 1115 516" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-      </svg>
-    ),
-    'spec2-left': (
-      <svg viewBox="0 0 372 511" className="w-full h-full" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M24.1372 18.8716C25.5845 18.5097 26.8226 13.7992 26.8226 15.2911C26.8226 18.9253 26.8048 22.1403 25.4799 25.585C20.944 37.3784 22.1007 52.6632 20.8551 65.1196C14.0677 132.993 11.869 203.684 21.4518 271.296C26.1418 304.386 31.5782 338.969 48.7531 368.268C65.6807 397.144 83.2488 410.751 113.351 424.362C147.103 439.624 179.337 441.453 215.992 440.325C242.486 439.51 269.288 434.681 295.359 440.474" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-        <path d="M273.304 374.396C261.596 358.298 282.197 376.932 286.731 380.363C305.211 394.348 330.169 411.378 342.825 430.937C348.749 440.093 367.724 467.705 346.704 471.069C315.049 476.134 284.841 485.126 254.506 495.237" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-      </svg>
-    ),
-    'spec2-right': (
-      <svg viewBox="0 0 372 511" className="w-full h-full" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M347.222 18.8716C345.775 18.5097 344.537 13.7992 344.537 15.2911C344.537 18.9253 344.555 22.1403 345.879 25.585C350.415 37.3784 349.259 52.6632 350.504 65.1196C357.292 132.993 359.49 203.684 349.908 271.296C345.218 304.386 339.781 338.969 322.606 368.268C305.679 397.144 288.111 410.751 258.008 424.362C224.257 439.624 192.022 441.453 155.368 440.325C128.873 439.51 102.071 434.681 76 440.474" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-        <path d="M98.0556 374.396C109.763 358.298 89.1623 376.932 84.6287 380.363C66.1481 394.348 41.1904 411.378 28.5344 430.937C22.61 440.093 3.63492 467.705 24.6555 471.069C56.3105 476.134 86.5185 485.126 116.853 495.237" stroke="currentColor" strokeWidth="30" strokeLinecap="round"/>
-      </svg>
-    ),
-    'spec3-left': (
-      <svg viewBox="0 0 740 111" className="w-full h-full" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M15 67.2037C227.98 67.2037 440.896 54.9199 653.757 54.9199" stroke="currentColor" strokeWidth="30" strokeLinecap="round" strokeLinejoin="round"/>
-        <path d="M681.267 15C688.451 20.1633 694.873 26.4372 701.525 32.249C706.61 36.6918 711.694 41.1377 716.827 45.525C718.809 47.2188 723.012 49.7164 724.261 52.2515C725.083 53.9181 723.077 55.0725 722.098 55.9688C713.986 63.3911 704.808 69.7851 696.057 76.404C687.806 82.6449 679.69 89.387 671 95.0101" stroke="currentColor" strokeWidth="30" strokeLinecap="round" strokeLinejoin="round"/>
-      </svg>
-    ),
-    'spec3-right': (
-      <svg viewBox="0 0 740 111" className="w-full h-full" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M724.454 67.2037C511.474 67.2037 298.558 54.9199 85.6969 54.9199" stroke="currentColor" strokeWidth="30" strokeLinecap="round" strokeLinejoin="round"/>
-        <path d="M58.1876 15C51.0038 20.1633 44.5812 26.4372 37.9294 32.249C32.8444 36.6918 27.7606 41.1377 22.6275 45.525C20.6458 47.2188 16.4422 49.7164 15.193 52.2515C14.3718 53.9181 16.377 55.0725 17.3565 55.9688C25.4683 63.3911 34.6465 69.7851 43.3971 76.404C51.6479 82.6449 59.7641 89.387 68.4543 95.0101" stroke="currentColor" strokeWidth="30" strokeLinecap="round" strokeLinejoin="round"/>
-      </svg>
-    ),
-  };
-
-  return (
-    <div className="text-orange-200 dark:text-slate-700" style={{ width, height }}>
-      {arrows[direction] || arrows['down-left']}
-    </div>
-  );
-}
