+++ src/components/link-library/LinkLibrary.module.css
@@ -1,1 +1,1 @@
-.page{min-height:100vh;background:#f8fafc;color:#0f172a;padding:48px 24px 96px}.header,.list,.empty,.message{max-width:896px;margin-left:auto;margin-right:auto}.header{display:flex;align-items:center;justify-content:space-between;gap:24px;margin-bottom:32px}.header h1{margin:0;font-family:var(--font-merriweather),Georgia,serif;font-size:30px;font-weight:700;letter-spacing:-.025em}.controls{display:flex;flex:1;justify-content:flex-end;gap:10px;max-width:500px}.search,.tagSelect{height:42px;border:1px solid #e2e8f0;border-radius:7px;background:#fff;color:#64748b}.search{display:flex;flex:1;align-items:center;gap:9px;padding:0 12px}.search input{width:100%;border:0;background:transparent;color:#0f172a;font:inherit;outline:0}.search input::placeholder{color:#94a3b8}.tagSelect{position:relative;display:block;min-width:160px}.tagSelect select{width:100%;height:100%;padding:0 36px 0 12px;border:0;appearance:none;background:transparent;color:#475569;font:inherit;outline:0;cursor:pointer}.selectIcon{position:absolute;top:50%;right:12px;pointer-events:none;transform:translateY(-50%)}.srOnly{position:absolute;width:1px;height:1px;padding:0;margin:-1px;overflow:hidden;clip:rect(0,0,0,0);white-space:nowrap;border:0}.list{border-top:1px solid #e2e8f0}.row{padding:20px 0;border-bottom:1px solid #e2e8f0}.row a{display:inline-block;color:#0f172a;text-decoration:none}.row a:hover h2{color:#475569}.row h2{margin:0;font-size:15px;font-weight:600;line-height:1.4;transition:color .15s}.row p{max-width:690px;margin:6px 0 0;color:#64748b;font-size:14px;line-height:1.55}.empty{min-height:260px;display:flex;align-items:center;justify-content:center;flex-direction:column;gap:10px;border-top:1px solid #e2e8f0;color:#64748b;text-align:center}.empty h2{margin:4px 0 0;font-size:16px;font-weight:600;color:#334155}.empty p{margin:0;font-size:14px}.message{margin-bottom:18px;padding:10px 12px;border:1px solid #fed7aa;border-radius:7px;background:#fff7ed;color:#9a3412;font-size:13px;display:flex;align-items:center;justify-content:space-between}.message button{display:grid;place-items:center;width:28px;height:28px;border:0;background:transparent;color:inherit;cursor:pointer}.spin{animation:spin 1s linear infinite}@keyframes spin{to{transform:rotate(360deg)}}:global(.dark) .page{background:#0f172a;color:#f8fafc}:global(.dark) .header h1,:global(.dark) .row a,:global(.dark) .empty h2{color:#f8fafc}:global(.dark) .row p,:global(.dark) .empty{color:#94a3b8}:global(.dark) .search,:global(.dark) .tagSelect{background:#1e293b;border-color:#334155;color:#cbd5e1}:global(.dark) .search input{color:#f8fafc}:global(.dark) .list,:global(.dark) .row,:global(.dark) .empty{border-color:#334155}:global(.dark) .row a:hover h2{color:#cbd5e1}@media(max-width:560px){.page{padding:36px 18px 72px}.header{align-items:stretch;flex-direction:column}.controls{max-width:none}.header h1{font-size:28px}.tagSelect{min-width:0}}
+.page{min-height:100vh;background:#f8fafc;color:#0f172a;padding:48px 24px 96px}.header,.list,.empty,.message{max-width:896px;margin-left:auto;margin-right:auto}.header{display:flex;align-items:center;justify-content:space-between;gap:24px;margin-bottom:32px}.header h1{margin:0;font-family:var(--font-merriweather),Georgia,serif;font-size:30px;font-weight:700;letter-spacing:-.025em}.controls{display:flex;flex:1;justify-content:flex-end;gap:10px;max-width:500px}.search,.tagSelect{height:42px;border:1px solid #e2e8f0;border-radius:7px;background:#fff;color:#64748b}.search{display:flex;flex:1;align-items:center;gap:9px;padding:0 12px}.search input{width:100%;border:0;background:transparent;color:#0f172a;font:inherit;outline:0}.search input::placeholder{color:#94a3b8}.tagSelect{position:relative;display:block;min-width:160px}.tagSelect select{width:100%;height:100%;padding:0 36px 0 12px;border:0;appearance:none;background:transparent;color:#475569;font:inherit;outline:0;cursor:pointer}.selectIcon{position:absolute;top:50%;right:12px;pointer-events:none;transform:translateY(-50%)}.srOnly{position:absolute;width:1px;height:1px;padding:0;margin:-1px;overflow:hidden;clip:rect(0,0,0,0);white-space:nowrap;border:0}.list{border-top:1px solid #e2e8f0}.row{border-bottom:1px solid #e2e8f0}.rowLink{display:flex;align-items:center;justify-content:space-between;gap:24px;padding:20px 0;color:#0f172a;text-decoration:none}.rowLink:hover h2{color:#475569}.row h2{margin:0;font-size:15px;font-weight:600;line-height:1.4;transition:color .15s}.row p{max-width:690px;margin:6px 0 0;color:#64748b;font-size:14px;line-height:1.55}.preview{width:120px;height:68px;flex:none;border:1px solid #e2e8f0;border-radius:5px;object-fit:cover}.empty{min-height:260px;display:flex;align-items:center;justify-content:center;flex-direction:column;gap:10px;border-top:1px solid #e2e8f0;color:#64748b;text-align:center}.empty h2{margin:4px 0 0;font-size:16px;font-weight:600;color:#334155}.empty p{margin:0;font-size:14px}.message{margin-bottom:18px;padding:10px 12px;border:1px solid #fed7aa;border-radius:7px;background:#fff7ed;color:#9a3412;font-size:13px;display:flex;align-items:center;justify-content:space-between}.message button{display:grid;place-items:center;width:28px;height:28px;border:0;background:transparent;color:inherit;cursor:pointer}.spin{animation:spin 1s linear infinite}@keyframes spin{to{transform:rotate(360deg)}}:global(.dark) .page{background:#0f172a;color:#f8fafc}:global(.dark) .header h1,:global(.dark) .rowLink,:global(.dark) .empty h2{color:#f8fafc}:global(.dark) .row p,:global(.dark) .empty{color:#94a3b8}:global(.dark) .search,:global(.dark) .tagSelect{background:#1e293b;border-color:#334155;color:#cbd5e1}:global(.dark) .search input{color:#f8fafc}:global(.dark) .list,:global(.dark) .row,:global(.dark) .empty{border-color:#334155}:global(.dark) .preview{border-color:#475569}:global(.dark) .rowLink:hover h2{color:#cbd5e1}@media(max-width:560px){.page{padding:36px 18px 72px}.header{align-items:stretch;flex-direction:column}.controls{max-width:none}.header h1{font-size:28px}.tagSelect{min-width:0}.rowLink{gap:16px}.preview{width:88px;height:56px}.row p{font-size:13px}}
